requirements.txt
Django==5.0.1
reportlab==4.1.0
Pillow==10.1.0